requirements.txt
faiss-cpu
numpy
pandas
sentence-transformers
streamlit
torch
gpt4all
langchain
langchain-community
transformers

